requirements.txt
openai
streamlit
pdfplumber
pytesseract
Pillow
python-dotenv
pandas
